jabatan views

+++ app/Repository/PegawaiRepository.php
@@ -1,9 +1,11 @@
 <?php
 namespace App\Repository;
 
+use Config\Database;
+use App\Helpers\ArrayHelper;
 use App\Models\PegawaiModel;
-use App\Entities\PegawaiEntities;
 use App\Repository\Repository;
+use App\Entities\PegawaiEntities;
 
 
 class PegawaiRepository extends Repository
@@ -32,7 +34,7 @@ class PegawaiRepository extends Repository
     /**
      * Entry Data Jabatan
      * @param {array} $object - data array jabatan
-     * @return {boolean}
+     * @return {void}
      */
     public function insert($object) {
         $item = new PegawaiEntities();
@@ -40,21 +42,27 @@ class PegawaiRepository extends Repository
         $item->nama = $object['nama'];
         $item->jabatan = $object['jabatan'];
         $item->golongan = $object['golongan'];
-        $item->foto = $object['foto'];
+        // $item->foto = $object['foto']; // just in case we can uncomment this field
 
         $model = new PegawaiModel();
-        return $model->insert($item);
+
+        if (!$this->isExistEmployee($id)) {
+            $model->insert($item);
+        }
     }
 
     /**
      * Update Data Pegawai
      * @param {integer} $id - nip
      * @param {array} $object - data array pegawai
-     * @return {boolean}
+     * @return {void}
      */
     public function update($id, $object) {
         $model = new PegawaiModel();
-        return $model->update($id, $object);
+
+        if ($this->isExistEmployee($id)) {
+            $model->update($id, $object);
+        }
     }
 
     /**
@@ -63,6 +71,74 @@ class PegawaiRepository extends Repository
      */
     public function delete($id) {
         $model = new PegawaiModel();
-        $model->delete($id);
+
+        if ($this->isExistEmployee($id)) {
+            $model->delete($id);
+        }
+    }
+
+    /**
+     * Is Exist Employee According NIP and Name
+     * @param {id} $id - employee ID
+     * @return {boolean}
+     */
+    public function isExistEmployee($id) {
+        $model = new PegawaiModel();
+        $employeeList = $model
+            ->where('nip', $id)
+            ->findAll();
+
+        return count($employeeList) >= 1;
+    }
+
+    /**
+     * Find Employee Formatted
+     * @return {mixed}
+     */
+    public function findEmployeeFormatted() {
+        $db = Database::connect();
+        $field = ArrayHelper::objectToFieldQuery([
+            'pegawai.nip' => 'pegawai_nip',
+            'pegawai.nama' => 'pegawai_nama',
+            'pegawai.jabatan' => 'pegawai_jabatan',
+            'pegawai.golongan' => 'pegawai_golongan',
+            'golongan.id_golongan' => 'golongan_id_golongan',
+            'golongan.golongan' => 'golongan_golongan',
+            'jabatan.id_jabatan' => 'jabatan_id_jabatan',
+            'jabatan.jabatan' => 'jabatan_jabatan',
+            'jabatan.jenjang_jabatan' => 'jabatan_jenjang_jabatan'
+        ]);
+        $response = $db
+            ->table('pegawai')
+            ->select($field)
+            ->join('jabatan', 'jabatan.id_jabatan = pegawai.jabatan')
+            ->join('golongan', 'golongan.id_golongan = pegawai.golongan')
+            ->get()
+            ->getResultArray();
+
+        return array_map(
+            function($value) {
+                $result = new \stdClass();
+                $result->nip = $value['pegawai_nip'];
+                $result->nama = $value['pegawai_nama'];
+                $result->jabatan = $value['pegawai_jabatan'];
+                $result->golongan = $value['pegawai_golongan'];
+
+                $golongan = new \stdClass();
+                $golongan->id_golongan = $value['golongan_id_golongan'];
+                $golongan->golongan = $value['golongan_golongan'];
+
+                $jabatan = new \stdClass();
+                $jabatan->id_jabatan = $value['jabatan_id_jabatan'];
+                $jabatan->jabatan = $value['jabatan_jabatan'];
+                $jabatan->jenjang_jabatan = $value['jabatan_jenjang_jabatan'];
+
+                $result->golongan = $golongan;
+                $result->jabatan = $jabatan;
+
+                return $result;
+            },
+            $response
+        );
     }
 }
